added in case insensitivity for XML tags as well as attribute values added in the ability to define constants in the definition xml by added the extenedelements tag to definitions and defining either a DefinitionVariable or DefinitionFile added in the ability to define runtime constants for a process at the point of creation
Both tweaks above now mean variables are obtained form the process variable container in the following order (Container->Runtime->Definition Constants), so during any process run new variables with the same name can be defined and they will be returned when accessing the variable.  If the variable name is not contained within the process variables, the runtime constants are checked, still not found, definition constants are checked, still not found, return null.