boundary drawing
completed the implementation of boundary event rendering, now need to run full tests with the new delegates as well as the boundary events to ensure that the system operates as expected.  Also need to implement a process step scheduled execution after a state load has occured to ensure that any delayed steps get properly executed.